Разделение проекта на два сервиса

+++ Core/entities/products/Product.cs
@@ -8,8 +8,8 @@ namespace Core.entities.products
         public string Name { get; set; }
 
         public string Description { get; set; }
-        public double Price { get; set; }
-        public uint Quantity { get; set; }
+        public float Price { get; set; }
+        public int Quantity { get; set; }
         public DateTime DatePrice { get; set; }
     }
 }
